requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.4.0
dj-database-url==1.2.0
python-dotenv==1.0.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
requests==2.31.0